+++ src/PluginProcessor.h
@@ -21,6 +21,7 @@ public:
     IntersectProcessor();
     ~IntersectProcessor() override;
 
+    using juce::AudioProcessor::processBlock;
     void prepareToPlay (double sampleRate, int samplesPerBlock) override;
     void releaseResources() override;
     void processBlock (juce::AudioBuffer<float>&, juce::MidiBuffer&) override;